clk_gate.c: 6: op: MISRA.ETYPE.ASSIGN.2012
- Rule 10.3, Essential type assignment violation: Expression value of essential type 'unsigned long int' assigned to object of essential type 'unsigned int'
- Fixed 2 violations (IDs: 56, 57) by adding explicit uint32_t casts to BIT() macro results

+++ plat/ti/k3low/common/drivers/pm/clock/clk_gate.c
@@ -20,9 +20,9 @@ static bool clk_gate_set_state(struct clk *clkp, bool enable)
 
 	v = readl(reg->reg);
 	if (enable) {
-		v |= BIT(reg->bit);
+		v |= (uint32_t)BIT(reg->bit);
 	} else {
-		v &= ~BIT(reg->bit);
+		v &= ~(uint32_t)BIT(reg->bit);
 	}
 	ti_clk_writel(v, reg->reg);
